+++ tomcat/src/main/java/com/mapping/DynamicControllerMapping.java
@@ -1,5 +1,9 @@
-package org.apache.web;
+package com.mapping;
 
+import com.controller.Controller;
+import com.controller.LoginController;
+import com.controller.RegisterController;
+import com.controller.RootController;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.coyote.http11.HttpMethod;
